Fixed some issues, added a few controllers and routes

+++ src/Backend/model/user.js
@@ -23,9 +23,8 @@ class UserDAO {
             if(JSON.stringify(check.rows)!=="[]"){
                 throw "error, user in system"
             }
-            const result = await Arres_db.query
+            await Arres_db.query
             ('INSERT INTO "User" ("email", "password", "first_name", "last_name", "isProfessor", "Concentration") VALUES($1,$2,$3,$4,$5,$6)', query);
-            console.log(result.rows)
             return JSON.stringify(query);
         } catch (error) {
             console.error("Email is already in system");
@@ -35,16 +34,16 @@ class UserDAO {
         }
     }
 
-    static get_all_user() {
-        Arres_db.connect();
-        Arres_db.query('SELECT * FROM "User"', (err, res) => {
-            if (!err) {
-                console.log(res.rows);
-            } else {
-                console.log(err);
-            }
+    static async get_all_user() {
+        try{
+            await Arres_db.connect();
+            const result = await Arres_db.query('SELECT * FROM "User"');
+            return JSON.stringify(result.rows);
+        } catch{
+            return false;
+        } finally {
             Arres_db.end();
-        })
+        }
     }
 
     static async get_user_by_id(user_id){
@@ -65,28 +64,45 @@ class UserDAO {
         }
     }
 
-    static check_user(email, password){
-        Arres_db.connect();
-        Arres_db.query('SELECT * FROM "User" where email = $1 AND password = $2', [JSON.parse(email), JSON.parse(password)], (err, res) => {
-            if (!err) {
-                console.log(res.rows);
-            } else {
-                return false;
+    static async check_user(email, password){
+        try{
+            await Arres_db.connect();
+            const result = await Arres_db.query('SELECT * FROM "User" where email = $1 AND password = $2', [JSON.parse(email), JSON.parse(password)])
+            if(JSON.stringify(result.rows)!=="[]"){
+                return JSON.stringify(result.rows);
+            } else{
+                throw "User not in system";
             }
+        }catch{
+            console.log("User not found!")
+        } finally {
             Arres_db.end();
-        })
+        }
+
     }
 
-    static delete_user(user_id){
-        Arres_db.connect();
-        Arres_db.query('DELETE FROM "User" where us_id = $1', [user_id], (err, res) => {
-            if (!err) {
-                console.log("Process Finished");
-            } else {
-                console.log(err);
-            }
-            Arres_db.end();
-        })
+    static async delete_user(user_id){
+        await Arres_db.connect();
+        const result = await Arres_db.query('SELECT * FROM "User" WHERE us_id=$1', [user_id]);
+        let json = JSON.stringify(result.rows);
+        if(json!=="[]"){
+            Arres_db.query('DELETE FROM "User" where us_id = $1', [user_id], (err, res) => {
+                if (!err) {
+                    const res = JSON.parse(json, (key, value) =>{
+                        if(key==="us_id"){
+                            console.log(JSON.stringify(value));
+                        }
+                    })
+                } else {
+                    console.log(err);
+                }
+                Arres_db.end();
+            })
+            return json;
+        }else{
+            console.log("User does not exist!")
+        }
+       Arres_db.end();
     }
 
 }
